Working /visit Command, Teleport to European Cities, PlayerMovementListener Initial
We make a visit command that teleports the player to the placed location. However, this leads to the discovery of a few bugs- an ECEF rotation issue still present with non-European cities and placing everything from session folder, which will be solved later when Python is merged with Java code. We also add an initial player movement listener for automatically loading new areas as the player moves.